HZN-1195: Move ElasticRule to its own test module
