+++ core/test-api/elasticsearch/pom.xml
@@ -9,54 +9,57 @@
   <groupId>org.opennms.core.test-api</groupId>
   <artifactId>org.opennms.core.test-api.elasticsearch</artifactId>
   <name>OpenNMS :: Core :: Test API :: Elasticsearch</name>
-  <packaging>bundle</packaging>
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.felix</groupId>
-        <artifactId>maven-bundle-plugin</artifactId>
-        <extensions>true</extensions>
-        <configuration>
-          <instructions>
-            <Bundle-RequiredExecutionEnvironment>JavaSE-1.8</Bundle-RequiredExecutionEnvironment>
-            <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
-            <Bundle-Version>${project.version}</Bundle-Version>
-          </instructions>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-  <dependencies>
-    <dependency>
-      <groupId>commons-io</groupId>
-      <artifactId>commons-io</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.camel</groupId>
-      <artifactId>camel-elasticsearch</artifactId>
-      <version>${camelVersion}</version>
-      <exclusions>
-        <exclusion>
-          <groupId>com.sun.xml.bind</groupId>
-          <artifactId>jaxb-impl</artifactId>
-        </exclusion>
-      </exclusions>
-      <!-- use the same version as your Camel core version -->
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.opennms.dependencies</groupId>
-      <artifactId>camel-test-dependencies</artifactId>
-      <type>pom</type>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.opennms</groupId>
-      <artifactId>opennms-icmp-api</artifactId>
-    </dependency>
-  </dependencies>
+  <packaging>jar</packaging>
+   <dependencies>
+     <dependency>
+       <groupId>org.slf4j</groupId>
+       <artifactId>slf4j-api</artifactId>
+     </dependency>
+     <dependency>
+       <groupId>org.elasticsearch</groupId>
+       <artifactId>elasticsearch</artifactId>
+       <version>5.6.2</version>
+     </dependency>
+     <dependency>
+       <groupId>org.elasticsearch.client</groupId>
+       <artifactId>transport</artifactId>
+       <version>5.6.2</version>
+       <exclusions>
+         <exclusion>
+           <groupId>commons-logging</groupId>
+           <artifactId>commons-logging</artifactId>
+         </exclusion>
+       </exclusions>
+     </dependency>
+     <dependency>
+       <groupId>org.elasticsearch.plugin</groupId>
+       <artifactId>transport-netty4-client</artifactId>
+       <version>5.6.2</version>
+     </dependency>
+     <dependency>
+       <groupId>io.netty</groupId>
+       <artifactId>netty-all</artifactId>
+       <version>4.1.17.Final</version>
+     </dependency>
+     <dependency>
+       <groupId>org.apache.logging.log4j</groupId>
+       <artifactId>log4j-api</artifactId>
+       <version>2.7</version>
+     </dependency>
+     <dependency>
+       <groupId>org.apache.logging.log4j</groupId>
+       <artifactId>log4j-core</artifactId>
+       <version>2.7</version>
+     </dependency>
+     <dependency>
+       <groupId>junit</groupId>
+       <artifactId>junit</artifactId>
+       <scope>compile</scope>
+     </dependency>
+     <dependency>
+       <groupId>org.hamcrest</groupId>
+       <artifactId>hamcrest-library</artifactId>
+       <scope>compile</scope>
+     </dependency>
+ </dependencies>
 </project>
