+++ core/test-api/elasticsearch/src/main/java/org/opennms/core/test/elastic/ElasticSearchServerConfig.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.flows.elastic.utils;
+package org.opennms.core.test.elastic;
 
 import java.util.Objects;
