+++ core/test-api/elasticsearch/src/main/java/org/opennms/core/test/elastic/EmbeddedElasticSearchServer.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.flows.elastic.utils;
+package org.opennms.core.test.elastic;
 
 import java.io.IOException;
 import java.util.Arrays;
