+++ core/test-api/elasticsearch/src/main/java/org/opennms/core/test/elastic/ExecutionTime.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.flows.elastic.utils;
+package org.opennms.core.test.elastic;
 
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
