+++ core/test-api/pom.xml
@@ -16,6 +16,7 @@
     <module>collection</module>
     <module>db</module>
     <module>dns</module>
+    <module>elasticsearch</module>
     <module>http</module>
     <module>kafka</module>
     <module>karaf</module>
